% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorEval.R
\name{computeClusterEvalStats}
\alias{computeClusterEvalStats}
\title{Calculate statistics for evaluating batch correction}
\usage{
computeClusterEvalStats(spe_object, foiColumn)
}
\arguments{
\item{spe_object}{A Spatial Experiment object.}

\item{foiColumn}{A column name indicating the factor of interest to be tested, can be biological factor or batch factor.}
}
\value{
A dataframe object containing the clustering evaluating statistics.
}
\description{
Calculate statistics for evaluating batch correction
}
\examples{
library(scater)
data("dkd_spe_subset")
spe <- scater::runPCA(dkd_spe_subset)
computeClusterEvalStats(spe, "SlideName")

}
