% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standR-package.R
\docType{data}
\name{dkd_spe_subset}
\alias{dkd_spe_subset}
\title{Description of the standR example datasets}
\format{
A spatial experiment object with 3000 rows and 70 samples:
}
\source{
\url{http://nanostring-public-share.s3-website-us-west-2.amazonaws.com/GeoScriptHub/KidneyDataset/}
}
\usage{
data("dkd_spe_subset")
}
\description{
standR-package has 1 datasets: \itemize{
\item dkd_spe_subset  Example subset of a GeoMX DSP WTA dataset,
}
}
\examples{
data(dkd_spe_subset)
}
\keyword{internal}
