% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plotting.R
\name{expand.grid.rmdup}
\alias{expand.grid.rmdup}
\title{Set up the grid for multiple-panel PCA plots}
\usage{
expand.grid.rmdup(x, y, include.equals = FALSE)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{include.equals}{Logical}
}
\description{
Set up the grid for multiple-panel PCA plots
}
