% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomx_import_data.R
\name{geomx_import_fun}
\alias{geomx_import_fun}
\title{Import geomx data into spe object}
\usage{
geomx_import_fun(
  dirPath,
  countFile,
  sampleAnnoFile,
  featureAnnoFile,
  hasNegProbe,
  NegProbeName,
  colnames.as.rownames,
  coord.colnames
)
}
\arguments{
\item{dirPath}{The folder path to all necessary tsv files.}

\item{countFile}{tsv file. Count matrix, with samples in columns and features/genes in rows. The first column is gene names/ids.}

\item{sampleAnnoFile}{tsv file. Sample annotations.}

\item{featureAnnoFile}{tsv file. Feature/Gene annotations.}

\item{hasNegProbe}{Logical. Indicating there are negative probe genes in the data.}

\item{NegProbeName}{Character. Name of negative probe genes.}

\item{colnames.as.rownames}{Vector of characters, length of 3. Column names used to capture gene names, sample names and gene names in countFile, sampleAnnoFile and featureAnnoFile, respectively.}

\item{coord.colnames}{Vector of characters, length of 2. Column names used to capture ROI coordinates.}
}
\value{
A Spatial Experiment object.
}
\description{
Import geomx data into spe object
}
