% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{geomx_normalize}
\alias{geomx_normalize}
\title{Perform normalisation to GeoMX data}
\usage{
geomx_normalize(spe_object, method = "TMM", log = TRUE)
}
\arguments{
\item{spe_object}{A spatial experiment object.}

\item{method}{Normalisation method to use.}

\item{log}{Log-transformed or not.}
}
\value{
A spatial experiment object.
}
\description{
Perform normalisation to GeoMX data
}
