% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plotting.R
\name{plot2dimensionPCA}
\alias{plot2dimensionPCA}
\title{Plot PCA with two dimensions}
\usage{
plot2dimensionPCA(
  spe_object,
  featureToPlot,
  featureToShape = NA,
  mainPCs = c("PC1", "PC2"),
  color = NA,
  legend.title = NA,
  font.size = 15,
  shape.size = 6,
  plotName = FALSE
)
}
\arguments{
\item{spe_object}{A spatial experiment object.}

\item{featureToPlot}{Feature to color by.}

\item{featureToShape}{Feature to shape by.}

\item{mainPCs}{Which PCs to plot.}

\item{color}{Color values.}

\item{legend.title}{Legend title.}

\item{font.size}{Font size.}

\item{shape.size}{Point size.}

\item{plotName}{if TRUE, plot the name of the samples.}
}
\value{
A ggplot object.
}
\description{
Plot PCA with two dimensions
}
