% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plotting.R
\name{plot2dUMAP}
\alias{plot2dUMAP}
\title{Plot UMAP}
\usage{
plot2dUMAP(spe, featureToPlot, point_size = 1, color = NA, emphasize = NULL)
}
\arguments{
\item{spe}{A spatial experiment object.}

\item{featureToPlot}{Feautre to color by.}

\item{point_size}{Point size.}

\item{color}{Colors.}

\item{emphasize}{Logical vector, indicating which samples are highlighted while the others are gray.}
}
\value{
A spatial experiment object.
}
\description{
Plot UMAP
}
