% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plotting.R
\name{plot_screePCA}
\alias{plot_screePCA}
\title{Plot the PCA scree plot.}
\usage{
plot_screePCA(
  spe_object,
  bar_color = "black",
  bar_fill = "royalblue",
  bar_width = 0.8,
  point_col = "tomato3",
  line_col = "tomato3",
  point_size = 2
)
}
\arguments{
\item{spe_object}{A spatial experiment object.}

\item{bar_color}{Color for bar.}

\item{bar_fill}{Fill for bar.}

\item{bar_width}{Bar width.}

\item{point_col}{Color for point.}

\item{line_col}{Color for line.}

\item{point_size}{Point size.}
}
\value{
A ggplot object.
}
\description{
Plot the PCA scree plot.
}
