% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQCs.R
\name{plotNEGpercentHist}
\alias{plotNEGpercentHist}
\title{Plot the distribution of the percentage of low expressed genes in samples.}
\usage{
plotNEGpercentHist(spe, hist_col, hist_fill, bin_num, text_size)
}
\arguments{
\item{spe}{A spatial experiment object.}

\item{hist_col}{Color for histogram.}

\item{hist_fill}{Fill for histogram.}

\item{bin_num}{Bin numbers for histogram.}

\item{text_size}{Text size.}
}
\value{
A ggplot object.
}
\description{
Plot the distribution of the percentage of low expressed genes in samples.
}
