% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plotting.R
\name{plotPairdimensionPCA}
\alias{plotPairdimensionPCA}
\title{Plot pair-wise PCA plots for multiple dimensions}
\usage{
plotPairdimensionPCA(
  spe_object,
  featureToPlot,
  featureToShape = NA,
  n_dimension = 3,
  color = NA,
  legend.title = NA,
  font.size = 15,
  shape.size = 6
)
}
\arguments{
\item{spe_object}{A spatial experiment object.}

\item{featureToPlot}{Feature to color by.}

\item{featureToShape}{Feature to shape by.}

\item{n_dimension}{The top n dimensions to be plotted}

\item{color}{Color values.}

\item{legend.title}{Legend title.}

\item{font.size}{Font size.}

\item{shape.size}{Dot size.}
}
\value{
A ggplot object.
}
\description{
Plot pair-wise PCA plots for multiple dimensions
}
