% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDR.R
\docType{methods}
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,DGEList-method}
\alias{plotPCA,ExpressionSet-method}
\alias{plotPCA,SummarizedExperiment-method}
\alias{plotPCA,SingleCellExperiment-method}
\alias{plotPCA,SpatialExperiment-method}
\alias{plotMDS,SummarizedExperiment-method}
\alias{plotMDS,SingleCellExperiment-method}
\alias{plotMDS,SpatialExperiment-method}
\title{Compute and plot the results of a PCA analysis on gene expression data}
\usage{
\S4method{plotPCA}{DGEList}(object, dims = c(1, 2), precomputed = NULL, rl = 1, ...)

\S4method{plotPCA}{ExpressionSet}(object, dims = c(1, 2), precomputed = NULL, rl = 1, ...)

\S4method{plotPCA}{SummarizedExperiment}(object, dims = c(1, 2), assay = 1, precomputed = NULL, rl = 1, ...)

\S4method{plotPCA}{SingleCellExperiment}(object, dims = c(1, 2), assay = 1, precomputed = NULL, rl = 1, ...)

\S4method{plotPCA}{SpatialExperiment}(object, dims = c(1, 2), assay = 1, precomputed = NULL, rl = 1, ...)

\S4method{plotMDS}{SummarizedExperiment}(object, dims = c(1, 2), precomputed = NULL, rl = 1, assay = 1, ...)

\S4method{plotMDS}{SingleCellExperiment}(object, dims = c(1, 2), precomputed = NULL, rl = 1, assay = 1, ...)

\S4method{plotMDS}{SpatialExperiment}(object, dims = c(1, 2), precomputed = NULL, rl = 1, assay = 1, ...)
}
\arguments{
\item{object}{a DGEList, SummarizedExperiment or ExpressionSet object
containing gene expression data.}

\item{dims}{a numeric, containing 2 values specifying the dimensions to plot.}

\item{precomputed}{a dimensional reduction results from \code{stats::prcomp}.
result in \code{reducedDims(object)} to plot.}

\item{rl}{a numeric, specifying the relative scale factor to apply to text on
the plot.}

\item{...}{aesthetic mappings to pass to \code{ggplot2::aes_string()}.}

\item{assay}{a numeric or character, specifying the assay to use (for
\code{SummarizedExperiment} and its derivative classes).}
}
\value{
a ggplot2 object
}
\description{
Compute and plot the results of a PCA analysis on gene expression data
}
\examples{
data("dkd_spe_subset")
plotPCA(dkd_spe_subset)

}
