% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRLE.R
\name{plotRLE}
\alias{plotRLE}
\alias{plotRLE,DGEList-method}
\alias{plotRLE,ExpressionSet-method}
\alias{plotRLE,SummarizedExperiment-method}
\title{Compute and plot relative log expression (RLE) values of gene expression data}
\usage{
plotRLE(edata, ordannots = c(), rl = 1, ...)

\S4method{plotRLE}{DGEList}(edata, ordannots = c(), rl = 1, ...)

\S4method{plotRLE}{ExpressionSet}(edata, ordannots = c(), rl = 1, ...)

\S4method{plotRLE}{SummarizedExperiment}(edata, ordannots = c(), rl = 1, ...)
}
\arguments{
\item{edata}{a DGEList, SummarizedExperiment or ExpressionSet object
containing gene expression data.}

\item{ordannots}{variables or computations to sort samples by (tidy style).}

\item{rl}{a numeric, specifying the relative scale factor to apply to text on
the plot.}

\item{...}{aesthetic mappings to pass to \code{ggplot2::aes_string()}.}
}
\value{
a ggplot2 object, containing the RLE plot.
}
\description{
Compute and plot relative log expression (RLE) values of gene expression data
}
\examples{
se = emtdata::cursons2018_se()
dge = emtdata::asDGEList(se)
plotRLE(dge, colour = Subline, lty = Treatment, lwd = 2, ordannots =
c(Subline, Treatment))

}
