% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQCs.R
\name{plotRmGenes}
\alias{plotRmGenes}
\title{Plot the removed genes with their expressions in all samples from a spatial experiment object after QC}
\usage{
plotRmGenes(spe, top_n, point_size, line_type, line_col, line_cex, text_size)
}
\arguments{
\item{spe}{A spatial experiment object.}

\item{top_n}{Integer. Indicating the top n genes will be plotted.}

\item{point_size}{Numeric. Point size.}

\item{line_type}{Character. Line types for ggplot.}

\item{line_col}{Color for line.}

\item{line_cex}{Cex for line.}

\item{text_size}{Text size.}
}
\value{
A ggplot object.
}
\description{
Plot the removed genes with their expressions in all samples from a spatial experiment object after QC
}
