% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSampleInfo.R
\name{plotSampleInfo}
\alias{plotSampleInfo}
\title{Plot the user-defined meta data using alluvium plot}
\usage{
plotSampleInfo(spe_object, column2plot, textsize = 3)
}
\arguments{
\item{spe_object}{A SpatialExperiment object.}

\item{column2plot}{Which columns to plot.}

\item{textsize}{text size.}
}
\value{
A ggplot object
}
\description{
Plot the user-defined meta data using alluvium plot
}
\examples{

data("dkd_spe_subset")
plotSampleInfo(dkd_spe_subset, column2plot = c("SlideName", "disease_status", "region"))

}
