% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomx_import_data.R
\name{readGeoMx}
\alias{readGeoMx}
\title{Import GeoMX DSP data into a saptial experiment object from file paths}
\usage{
readGeoMx(
  dirPath,
  countFile,
  sampleAnnoFile,
  featureAnnoFile = NA,
  hasNegProbe = TRUE,
  NegProbeName = "NegProbe-WTX",
  colnames.as.rownames = c("TargetName", "SegmentDisplayName", "TargetName"),
  coord.colnames = c("ROICoordinateX", "ROICoordinateY")
)
}
\arguments{
\item{dirPath}{The folder path to all necessary tsv files.}

\item{countFile}{tsv file. Count matrix, with samples in columns and features/genes in rows. The first column is gene names/ids.}

\item{sampleAnnoFile}{tsv file. Sample annotations.}

\item{featureAnnoFile}{tsv file. Feature/Gene annotations.}

\item{hasNegProbe}{Logical. Default is TRUE, indicating there are negative probe genes in the data.}

\item{NegProbeName}{Character. Name of negative probe genes, default is NegProbe-WTX.}

\item{colnames.as.rownames}{Vector of characters, length of 3. Column names used to capture gene names, sample names and gene names in countFile, sampleAnnoFile and featureAnnoFile, respectively.}

\item{coord.colnames}{Vector of characters, length of 2. Column names used to capture ROI coordinates.}
}
\value{
A SpatialExperiment object.
}
\description{
Import GeoMX DSP data into a saptial experiment object from file paths
}
\examples{
dirPath <- system.file("extdata", package = "standR")
countFile <- "dkd_subset_TargetCountMatrix.txt"
sampleAnnoFile <- "dkd_subset_Sample_Annotations.txt"

spe <- readGeoMx(dirPath, countFile, sampleAnnoFile, hasNegProbe = FALSE)

}
