% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomx_import_data.R
\name{readGeoMxFromNanoStringGeoMxSet}
\alias{readGeoMxFromNanoStringGeoMxSet}
\title{Import GeoMX DSP data from a NanoStringGeoMxSet-like object}
\usage{
readGeoMxFromNanoStringGeoMxSet(
  geomxSet,
  assay2use = "exprs",
  sampleIDCol = "SegmentDisplayName",
  featureIDCol = "TargetName",
  rmNegProbe = TRUE,
  NegProbeName = "NegProbe-WTX",
  coord.colnames = c("ROICoordinateX", "ROICoordinateY")
)
}
\arguments{
\item{geomxSet}{A \code{Biobase::ExpressionSet}-derived object.}

\item{assay2use}{Name of the assay data element to use. Default is \code{exprs}.}

\item{sampleIDCol}{Column in the sample annotation used as sample IDs. If the
column is absent it is created from the assay column names.}

\item{featureIDCol}{Column in the feature annotation used as feature IDs. If
duplicated, counts are summed by this column.}

\item{rmNegProbe}{Logical. Default is TRUE, indicating negative probe genes
are removed from the assay and stored in metadata.}

\item{NegProbeName}{Character. Name of negative probe genes.}

\item{coord.colnames}{Vector of characters, length of 2. Column names used to
capture ROI coordinates.}
}
\value{
A SpatialExperiment object.
}
\description{
Converts a \code{Biobase::ExpressionSet} object, including objects returned by
\code{GeomxTools::readNanoStringGeoMxSet()}, into a \code{SpatialExperiment} object.
}
\examples{
\dontrun{
library(GeomxTools)
datadir <- system.file("extdata", "DSP_NGS_Example_Data", package = "GeomxTools")
dccFiles <- dir(datadir, pattern = ".dcc$", full.names = TRUE)
pkcFiles <- unzip(zipfile = file.path(datadir, "pkcs.zip"))
sampleAnnotationFile <- file.path(datadir, "annotations.xlsx")

dccSet <- readNanoStringGeoMxSet(
  dccFiles = dccFiles,
  pkcFiles = pkcFiles,
  phenoDataFile = sampleAnnotationFile,
  phenoDataSheet = "CW005",
  phenoDataDccColName = "Sample_ID"
)
spe <- readGeoMxFromNanoStringGeoMxSet(dccSet)
}
}
