% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrection.R
\name{run_combat}
\alias{run_combat}
\title{Combat normalisation}
\usage{
run_combat(spe_object, n_assay, batch, bio_factor)
}
\arguments{
\item{spe_object}{A Spatial Experiment object.}

\item{n_assay}{Integer, choose the assay from spe_object.}

\item{batch}{A vector indicating batches.}

\item{bio_factor}{A vector indicating biology.}
}
\value{
A Spatial Experiment object.
}
\description{
Combat normalisation
}
