% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrection.R
\name{run_limmarmbatch}
\alias{run_limmarmbatch}
\title{Limma batch correction}
\usage{
run_limmarmbatch(spe_object, n_assay, batch)
}
\arguments{
\item{spe_object}{A Spatial Experiment object.}

\item{n_assay}{Integer, choose the assay from spe_object.}

\item{batch}{A vector indicating batches.}
}
\value{
A Spatial Experiment object.
}
\description{
Limma batch correction
}
