% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrection.R
\name{run_ruv4}
\alias{run_ruv4}
\title{RUV4 batch correction}
\usage{
run_ruv4(spe_object, k, factors, negctrlGenes)
}
\arguments{
\item{spe_object}{A Spatial Experiment object.}

\item{k}{The number of unwanted factors to use. Can be 0.}

\item{factors}{Column name(s) to indicate the factors of interest.}

\item{negctrlGenes}{Negative control genes.}
}
\value{
A Spatial Experiment object.
}
\description{
RUV4 batch correction
}
