% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrection.R
\name{runCombat}
\alias{runCombat}
\title{Combat batch correction}
\usage{
runCombat(spe_object, n_assay = 2, batch, bio_factor)
}
\arguments{
\item{spe_object}{A Spatial Experiment object.}

\item{n_assay}{Integer, choose the assay from spe_object.}

\item{batch}{A vector indicating batches.}

\item{bio_factor}{A vector indicating biology.}
}
\value{
A Spatial Experiment object.
}
\description{
Combat batch correction
}
\examples{
data("dkd_spe_subset")
spe_combat <- runCombat(dkd_spe_subset,
                    batch = SummarizedExperiment::colData(dkd_spe_subset)$SlideName,
                    bio_factor = SummarizedExperiment::colData(dkd_spe_subset)$region)
SummarizedExperiment::assay(spe_combat, 2)

}
