% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrection.R
\name{runLimmarmb}
\alias{runLimmarmb}
\title{Limma batch correction}
\usage{
runLimmarmb(spe_object, n_assay = 2, batch)
}
\arguments{
\item{spe_object}{A Spatial Experiment object.}

\item{n_assay}{Integer, choose the assay from spe_object.}

\item{batch}{A vector indicating batches.}
}
\value{
A Spatial Experiment object.
}
\description{
Limma batch correction
}
\examples{
data("dkd_spe_subset")
spe_limmarmb <- runLimmarmb(dkd_spe_subset,
                    batch = SummarizedExperiment::colData(dkd_spe_subset)$SlideName)
SummarizedExperiment::assay(spe_limmarmb, 2)

}
