% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrection.R
\name{runRUV4}
\alias{runRUV4}
\title{RUV4 batch correction}
\usage{
runRUV4(spe_object, k, factors, negctrlGenes)
}
\arguments{
\item{spe_object}{A Spatial Experiment object.}

\item{k}{The number of unwanted factors to use. Can be 0.}

\item{factors}{Column name(s) to indicate the factors of interest.}

\item{negctrlGenes}{Negative control genes.}
}
\value{
A Spatial Experiment object, containing the ruv4-normalised count and normalisation factor.
}
\description{
RUV4 batch correction
}
\examples{
data("dkd_spe_subset")
spe <- findNCGs(dkd_spe_subset, top_n = 100)
spe_ruv <- runRUV4(spe, k = 3,
                  factors = c("disease_status","region"),
                  negctrlGenes = S4Vectors::metadata(spe)$negGenes)

}
