% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{spe2dge}
\alias{spe2dge}
\title{Transfer SpatialExperiment object into DGEList object for DE analysis}
\usage{
spe2dge(spe)
}
\arguments{
\item{spe}{SpatialExperiment object.}
}
\value{
A DGEList.
}
\description{
Transfer SpatialExperiment object into DGEList object for DE analysis
}
\examples{
data("dkd_spe_subset")


spe_tmm <- geomxNorm(dkd_spe_subset, method = "TMM")
dge <- spe2dge(spe_tmm)

}
