% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrection.R
\name{chooseNegCTRLgenes}
\alias{chooseNegCTRLgenes}
\title{Get negative control genes from the data}
\usage{
chooseNegCTRLgenes(
  spe_object,
  n_assay = 2,
  batch_name = "SlideName",
  top_n = 200
)
}
\arguments{
\item{spe_object}{A Spatial Experiment object.}

\item{n_assay}{Integer to indicate the nth count table in the assay(spe) to be used.}

\item{batch_name}{Column name indicating batches.}

\item{top_n}{Integer indicate how many genes to be included as negative control genes.}
}
\value{
A Spatial Experiment object.
}
\description{
Get negative control genes from the data
}
