% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorEval.R
\name{computeClusterEvalStats}
\alias{computeClusterEvalStats}
\title{Calculate statistics for evaluating batch correction}
\usage{
computeClusterEvalStats(spe_object, foiColumn)
}
\arguments{
\item{spe_object}{A Spatial Experiment object.}

\item{foiColumn}{Factors of interest.}
}
\value{
A Spatial Experiment object.
}
\description{
Calculate statistics for evaluating batch correction
}
