% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dkd_spe_subset}
\alias{dkd_spe_subset}
\title{A Spatial Experiment object containing a subset of DKD WTA data.}
\format{
A spatial experiment object with 3000 rows and 70 samples:
}
\source{
\url{http://nanostring-public-share.s3-website-us-west-2.amazonaws.com/GeoScriptHub/KidneyDataset/}
}
\usage{
data("dkd_spe_subset")
}
\description{
A dataset containing 3000 genes and 70 ROIs
}
\keyword{datasets}
