% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomx_import_data.R
\name{geomx_import_from_dge}
\alias{geomx_import_from_dge}
\title{Import GeoMX DSP data into a saptial experiment object from DGEList object}
\usage{
geomx_import_from_dge(dge_object, spatialCoord = NULL)
}
\arguments{
\item{dge_object}{a DGEList object.}

\item{spatialCoord}{a matrix with coordinates of samples, rowname must be cosistent with the colnames of dge_object.}
}
\value{
A SpatialExperiment object.
}
\description{
Import GeoMX DSP data into a saptial experiment object from DGEList object
}
