% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrection.R
\name{geomxBatchCorrection}
\alias{geomxBatchCorrection}
\title{Batch correction for GeoMX data}
\usage{
geomxBatchCorrection(
  spe,
  k,
  factors,
  NCGs,
  n_assay = 2,
  batch,
  covariates = NULL,
  design = matrix(1, ncol(spe), 1),
  method = c("RUV4", "Limma")
)
}
\arguments{
\item{spe}{A Spatial Experiment object.}

\item{k}{The number of unwanted factors to use. Can be 0. This is required for the RUV4 method.}

\item{factors}{Column name(s) to indicate the factors of interest. This is required for the RUV4 method.}

\item{NCGs}{Negative control genes. This is required for the RUV4 method.}

\item{n_assay}{Integer to indicate the nth count table in the assay(spe) to be used.}

\item{batch}{A vector indicating batches. This is required for the Limma mehtod.}

\item{covariates}{A matrix or vector of numeric covariates to be adjusted for.}

\item{design}{A design matrix relating to treatment conditions to be preserved, can be generated using \code{stats::model.matrix} function with all biological factors included.}

\item{method}{Can be either RUV4 or Limma, by default is RUV4.}
}
\value{
A Spatial Experiment object, containing the ruv4-normalised count and normalisation factor.
}
\description{
Batch correction for GeoMX data
}
\examples{
data("dkd_spe_subset")
spe <- findNCGs(dkd_spe_subset, top_n = 100)
spe_ruv <- geomxBatchCorrection(spe, k = 3,
                  factors = c("disease_status","region"),
                  NCGs = S4Vectors::metadata(spe)$NCGs)

}
