% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{geomxNorm}
\alias{geomxNorm}
\title{Perform normalization to GeoMX data}
\usage{
geomxNorm(
  spe_object,
  method = c("TMM", "RPKM", "TPM", "CPM", "upperquartile", "sizefactor"),
  log = TRUE
)
}
\arguments{
\item{spe_object}{A SpatialExperiment object.}

\item{method}{Normalization method to use. Options: TMM, RPKM, TPM, CPM, upperquartile, sizefactor. RPKM and TPM require gene length information, which should be added into rowData(spe). Note that TMM here is TMM + CPM.}

\item{log}{Log-transformed or not.}
}
\value{
A SpatialExperiment object, with the second assay being the normalized count matrix.
}
\description{
Perform normalization to GeoMX data
}
\examples{
data("dkd_spe_subset")

spe_tmm <- geomxNorm(dkd_spe_subset, method = "TMM")
head(assay(spe_tmm, "logcounts"))
spe_upq <- geomxNorm(dkd_spe_subset, method = "upperquartile")
head(assay(spe_upq, "logcounts"))
spe_deseqnorm <- geomxNorm(dkd_spe_subset, method = "sizefactor")
head(assay(spe_deseqnorm, "logcounts"))

}
\references{
Robinson, M. D., McCarthy, D. J., & Smyth, G. K. (2010). edgeR: a Bioconductor package for differential expression analysis of digital gene expression data. Bioinformatics, 26(1), 139-140.

Love, M., Anders, S., & Huber, W. (2014). Differential analysis of count data–the DESeq2 package. Genome Biol, 15(550), 10-1186.
}
