% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorEval.R
\name{plotClusterEvalStats}
\alias{plotClusterEvalStats}
\title{Plot the evaluation statistics}
\usage{
plotClusterEvalStats(
  spe_list,
  bio_feature_name,
  batch_feature_name,
  data_names
)
}
\arguments{
\item{spe_list}{A Spatial Experiment object.}

\item{bio_feature_name}{Biological variation name.}

\item{batch_feature_name}{Batch variation name.}

\item{data_names}{Data names.}
}
\value{
A ggplot object.
}
\description{
Plot the evaluation statistics
}
