% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetadata.R
\name{plotMetadata}
\alias{plotMetadata}
\title{Plot the meta data using alluvium plot}
\usage{
plotMetadata(spe_object, column2plot, textsize = 3)
}
\arguments{
\item{spe_object}{A spatial experiment object.}

\item{column2plot}{Which columns to plot.}

\item{textsize}{text size.}
}
\value{
A ggplot object
}
\description{
Plot the meta data using alluvium plot
}
