% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plotting.R
\name{plotPCAbiplot}
\alias{plotPCAbiplot}
\title{Plot PCA bi plot}
\usage{
plotPCAbiplot(
  spe_object,
  featureToPlot,
  featureToShape = NA,
  n_loadings = 10,
  mainPCs = c("PC1", "PC2"),
  color = NA,
  legend.title = NA,
  font.size = 15,
  shape.size = 6
)
}
\arguments{
\item{spe_object}{A spatial experiment object.}

\item{featureToPlot}{Feautre to color by.}

\item{featureToShape}{Feature to shape by.}

\item{n_loadings}{Plot the top n gene loadings}

\item{mainPCs}{PCs to plot.}

\item{color}{Color values}

\item{legend.title}{Legend title.}

\item{font.size}{Font size.}

\item{shape.size}{Point size.}
}
\value{
A ggplot object.
}
\description{
Plot PCA bi plot
}
