% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRLE.R
\name{plotRLE}
\alias{plotRLE}
\alias{plotRLE,DGEList-method}
\alias{plotRLE,ExpressionSet-method}
\alias{plotRLE,SummarizedExperiment-method}
\title{Compute and plot relative log expression (RLE) values of gene expression data}
\usage{
plotRLE(object, ordannots = c(), ...)

\S4method{plotRLE}{DGEList}(object, ordannots = c(), ...)

\S4method{plotRLE}{ExpressionSet}(object, ordannots = c(), ...)

\S4method{plotRLE}{SummarizedExperiment}(object, ordannots, assay = 1, ...)
}
\arguments{
\item{object}{a DGEList, SummarizedExperiment or ExpressionSet object
containing gene expression data.}

\item{ordannots}{variables or computations to sort samples by (tidy style).}

\item{...}{aesthetic mappings to pass to \code{ggplot2::aes_string()}.}

\item{assay}{a numeric or character, specifying the assay to use (for
\code{SummarizedExperiment} and its derivative classes).}
}
\value{
a ggplot2 object, containing the RLE plot.
}
\description{
Compute and plot relative log expression (RLE) values of gene expression data
}
\examples{
se <- emtdata::cursons2018_se()
dge <- emtdata::asDGEList(se)
plotRLE(dge,
  colour = Subline, lty = Treatment, lwd = 2, ordannots =
    c(Subline, Treatment)
)

}
