% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRLE.R
\name{plotRLE}
\alias{plotRLE}
\title{Plot Relative log expression}
\usage{
plotRLE(
  spe_object,
  n_assay = 1,
  colorby = NA,
  sortby = colorby,
  font_size = 20,
  line_cor = "red"
)
}
\arguments{
\item{spe_object}{A spatial experiment object}

\item{n_assay}{integer. the nth assay in the spe object to plot.}

\item{colorby}{feature to color by.}

\item{sortby}{feature to sort by.}

\item{font_size}{font size.}

\item{line_cor}{line color.}
}
\value{
A ggplot object
}
\description{
Plot Relative log expression
}
