% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomx_import_data.R
\name{readGeoMx}
\alias{readGeoMx}
\title{Import GeoMX DSP data into a saptial experiment object from file paths}
\usage{
readGeoMx(
  countFile,
  sampleAnnoFile,
  featureAnnoFile = NA,
  hasNegProbe = TRUE,
  NegProbeName = "NegProbe-WTX",
  colnames.as.rownames = c("TargetName", "SegmentDisplayName", "TargetName"),
  coord.colnames = c("ROICoordinateX", "ROICoordinateY")
)
}
\arguments{
\item{countFile}{tsv file. Count matrix, with samples in columns and features/genes in rows. The first column is gene names/ids.}

\item{sampleAnnoFile}{tsv file. Sample annotations.}

\item{featureAnnoFile}{tsv file. Feature/Gene annotations.}

\item{hasNegProbe}{Logical. Default is TRUE, indicating there are negative probe genes in the data.}

\item{NegProbeName}{Character. Name of negative probe genes, default is NegProbe-WTX.}

\item{colnames.as.rownames}{Vector of characters, length of 3. Column names used to capture gene names, sample names and gene names in countFile, sampleAnnoFile and featureAnnoFile, respectively.}

\item{coord.colnames}{Vector of characters, length of 2. Column names used to capture ROI coordinates.}
}
\value{
A SpatialExperiment object.
}
\description{
Import GeoMX DSP data into a saptial experiment object from file paths
}
\examples{
url <- "http://nanostring-public-share.s3-website-us-west-2.amazonaws.com/"
countFile <- paste0(url,"GeoScriptHub/KidneyDataset/Kidney_Raw_TargetCountMatrix.txt")
sampleAnnoFile <- paste0(url,"GeoScriptHub/KidneyDataset/Kidney_Sample_Annotations.txt")

spe <- readGeoMx(countFile, sampleAnnoFile, hasNegProbe = FALSE)

}
