% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomx_import_data.R
\name{readGeoMxFromNanoStringGeoMxSet}
\alias{readGeoMxFromNanoStringGeoMxSet}
\title{Import GeoMX DSP data from a NanoStringGeoMxSet-like object}
\usage{
readGeoMxFromNanoStringGeoMxSet(
  geomxSet,
  assay2use = "exprs",
  sampleIDCol = "SegmentDisplayName",
  featureIDCol = "TargetName",
  rmNegProbe = TRUE,
  NegProbeName = "NegProbe-WTX",
  coord.colnames = c("ROICoordinateX", "ROICoordinateY")
)
}
\arguments{
\item{geomxSet}{A \code{Biobase::ExpressionSet}-derived object.}

\item{assay2use}{Name of the assay data element to use. Default is \code{exprs}.}

\item{sampleIDCol}{Column in the sample annotation used as sample IDs. If the
column is absent it is created from the assay column names.}

\item{featureIDCol}{Column in the feature annotation used as feature IDs. If
duplicated, counts are summed by this column.}

\item{rmNegProbe}{Logical. Default is TRUE, indicating negative probe genes
are removed from the assay and stored in metadata.}

\item{NegProbeName}{Character. Name of negative probe genes.}

\item{coord.colnames}{Vector of characters, length of 2. Column names used to
capture ROI coordinates.}
}
\value{
A SpatialExperiment object.
}
\description{
Converts a \code{Biobase::ExpressionSet}-derived object into a
\code{SpatialExperiment} object.
}
\examples{
expr <- matrix(c(10, 1), nrow = 2,
               dimnames = list(c("GeneA", "NegProbe-WTX"), "ROI1"))
pheno <- data.frame(ROICoordinateX = 1, ROICoordinateY = 1, row.names = "ROI1")
feature <- data.frame(TargetName = rownames(expr), row.names = rownames(expr))
eset <- Biobase::ExpressionSet(
  assayData = expr,
  phenoData = Biobase::AnnotatedDataFrame(pheno),
  featureData = Biobase::AnnotatedDataFrame(feature)
)
spe <- readGeoMxFromNanoStringGeoMxSet(eset)
}
