% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{spe2dge}
\alias{spe2dge}
\title{Transfer SpatialExperiment object into DGEList object for DE analysis}
\usage{
spe2dge(spe, norm = NULL)
}
\arguments{
\item{spe}{SpatialExperiment object.}

\item{norm}{Prior normalization method to be used. Norm factor calculated from the previous normalization step to be used in the DGEList.}
}
\value{
A DGEList.
}
\description{
Transfer SpatialExperiment object into DGEList object for DE analysis
}
\examples{
data("dkd_spe_subset")

dge <- spe2dge(dkd_spe_subset)

spe_tmm <- geomxNorm(dkd_spe_subset, method = "TMM")
dge <- spe2dge(spe_tmm, norm = "TMM")

}
